fix case of docker icons
filename matching works if the icon is indexed as "dockerfile" instead
of "Dockerfile". This should probably also be applied to all other
filename matches, but I was just interested in this one case ;)

Additionally, I've also added an entry for "docker-compose.yaml" in
addition to just "docker-compose", since the use of the explicit
extension seems pretty common and is also supported by docker-compose.

fixes #13